Move fixedmorton32 to barnes_hut.h
