--- dyn/morton.h
@@ -1,68 +1,0 @@
-#ifndef GRASS_MORTON_H
-#define GRASS_MORTON_H
-
-#include <array>
-#include <bit>
-#include <complex>
-#include <cstdint>
-#include <optional>
-
-namespace dyn {
-
-namespace morton::detail {
-
-/// @brief Interleave the bits of two 32-bit words (re, im) so that the word
-/// im is placed at the odd-numbered bits (including the most significant bit)
-/// while re is placed at the even-numbered bits (including the least
-/// significant bit). Bits are numbered from 0 (LSB) to 61 (MSB) inclusive.
-constexpr uint64_t interleave32(uint32_t re, uint32_t im) {
-  // Modification of "Interleave by Binary Magic Numbers"
-  // (http://graphics.stanford.edu/~seander/bithacks.html#InterleaveBMN)
-  struct Help {
-    uint64_t mask, shift;
-  };
-
-  // 32-bit words. Hence, 5.
-  // 0101..., 00110011...,
-  // 00001111..., 0000000011111111...,
-  // including until the entire word is like 0...0_1...1_0...0_1...1.
-  // (Reverse order, however.)
-  std::array<Help, 5> constexpr H{
-      Help{0x0000ffff0000ffff, 16}, Help{0x00ff00ff00ff00ff, 8},
-      Help{0x0f0f0f0f0f0f0f0f, 4}, Help{0x3333333333333333, 2},
-      Help{0x5555555555555555, 1}};
-
-  // Zero-extend each word.
-  std::array<uint64_t, 2> W{re, im};
-  for (auto &&w : W)
-    // Spread the bits out with zeros in between.
-    // (ex) w = 0b1011 -> 0b01'00'01'01
-    for (auto &&h : H)
-      w = (w | (w << h.shift)) & h.mask;
-
-  // Imaginary first.
-  return W[0] | (W[1] << 1);
-}
-} // namespace morton::detail
-
-/// @brief Compute the Morton (Z) code of a complex number xy assuming a squared
-/// grid by pre-multiplying the factor `Precision` to each component of xy.
-/// If either component is not a finite number after the scaling, the result
-/// is {} (no value).
-template <uint32_t Precision = 512>
-std::optional<uint64_t> fixedmorton32(std::complex<float> xy) {
-  xy *= float(Precision);
-  if (std::abs(xy.real()) < float(INT32_MAX) &&
-      std::abs(xy.imag()) < float(INT32_MAX)) {
-    // unsigned.
-    auto sgn = 0x8000'0000u;
-    auto x = int32_t(xy.real()) ^ sgn; // promoted to unsigned.
-    auto y = int32_t(xy.imag()) ^ sgn; // (ditto).
-    return morton::detail::interleave32(x, y);
-  }
-  return {};
-}
-
-} // namespace dyn
-
-#endif // GRASS_MORTON_H
